requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
simplejwt==2.0.1
sqlparse==0.4.3
typing==3.7.4.3
tzdata==2022.6
